// generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const account_type = {
  real: 'real',
  nominal: 'nominal',
  allocation: 'allocation'
} as const

export type account_type = (typeof account_type)[keyof typeof account_type]


export const asset_type = {
  rupees: 'rupees',
  mf: 'mf',
  etf: 'etf',
  shares: 'shares',
  other: 'other'
} as const

export type asset_type = (typeof asset_type)[keyof typeof asset_type]